refactor: rename section components and folder
Renamed components to include "Section" suffix (e.g., about.jsx to aboutSection.jsx) for better clarity. Renamed `section` folder to `layout` folder to better reflect its purpose within the project structure.
